requirements.txt
annotated-types==0.7.0
attrs==24.2.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.32.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
docstring_parser==0.16
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.72.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.0
grpcio-status==1.68.0
httplib2==0.22.0
idna==3.10
isodate==0.7.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
msrest==0.7.1
numpy==2.1.3
oauthlib==3.2.2
packaging==24.2
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.21.0
rsa==4.9
shapely==2.0.6
six==1.16.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
